group CRUD requires user acess

+++ src/controllers/GroupController.js
@@ -1,4 +1,5 @@
 const Group = require('../models/Group');
+const Person = require('../models/Person');
 
 module.exports = {
     async index(req, res){
@@ -14,10 +15,24 @@ module.exports = {
         return res.json(groups);
     },
     async store(req, res) {
+        const { userid } = req.headers;
+
+        if(!userid)
+            return res.json({error: "user not found"});
+
+        let person;
+
+        // try to find the person
+        try {
+            person = await Person.findById(userid);
+        } catch {
+            return res.json({error: "user not found"});
+        }
+
         const { name, category, media, url } = req.body;
 
+        // check if the group already exists
         const groupExists = await Group.findOne({url});
-
         if(groupExists) 
             return res.json(groupExists);
 
@@ -28,19 +43,57 @@ module.exports = {
             url
         });
 
+        //save the group
+        person.groups.push(group._id);
+        person.save();
+
         return res.json(group);
     },
     async update(req, res) {
-        const { id } = req.params;
-        const group = await Group.findByIdAndUpdate(id, req.body, { new: true });
+        const { userid } = req.headers;
 
-        return res.json(group);
+        if(!userid)
+            return res.json({error: "user not found"});
+
+        let person;
+
+        try {
+            person = await Person.findById(userid);
+        } catch {
+            return res.json({error: "user not found"});
+        }
+
+        const { id } = req.params;
+
+        if(person.groups.includes(id)) {
+            const group = await Group.findByIdAndUpdate(id, req.body, { new: true });
+            return res.json(group);
+        } 
+
+        return res.json({error: "group doesn't exists"});
+        
     },
     async delete(req, res) {
+        const { userid } = req.headers;
+
+        if(!userid)
+            return res.json({error: "user not found"});
+
+        let person;
+
+        try {
+            person = await Person.findById(userid);
+        } catch {
+            return res.json({error: "user not found"});
+        }
+
         const { id } = req.params;
 
-        await Group.findByIdAndDelete(id);
+        if(person.groups.includes(id)) {
+            await Group.findByIdAndDelete(id);
+            return res.json({ok: true});
+        } 
 
-        return res.json({ok: true});
+        return res.json({error: "group doesn't exists"});
     }
 }
